Move other auth methods to _auth
