+++ pymysql/_auth.py
@@ -9,10 +9,103 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization, hashes
 from cryptography.hazmat.primitives.asymmetric import padding
 
+from functools import partial
 import hashlib
+import struct
 
 
 DEBUG = True
+SCRAMBLE_LENGTH = 20
+sha1_new = partial(hashlib.new, 'sha1')
+
+
+# mysql_native_password
+# https://dev.mysql.com/doc/internals/en/secure-password-authentication.html#packet-Authentication::Native41
+
+
+def scramble_native_password(password, message):
+    """Scramble used for mysql_native_password"""
+    if not password:
+        return b''
+
+    stage1 = sha1_new(password).digest()
+    stage2 = sha1_new(stage1).digest()
+    s = sha1_new()
+    s.update(message[:SCRAMBLE_LENGTH])
+    s.update(stage2)
+    result = s.digest()
+    return _my_crypt(result, stage1)
+
+
+def _my_crypt(message1, message2):
+    length = len(message1)
+    result = b''
+    for i in range_type(length):
+        x = (
+            struct.unpack('B', message1[i:i + 1])[0] ^
+            struct.unpack('B', message2[i:i + 1])[0]
+        )
+        result += struct.pack('B', x)
+    return result
+
+
+# old_passwords support ported from libmysql/password.c
+# https://dev.mysql.com/doc/internals/en/old-password-authentication.html
+
+SCRAMBLE_LENGTH_323 = 8
+
+
+class RandStruct_323(object):
+
+    def __init__(self, seed1, seed2):
+        self.max_value = 0x3FFFFFFF
+        self.seed1 = seed1 % self.max_value
+        self.seed2 = seed2 % self.max_value
+
+    def my_rnd(self):
+        self.seed1 = (self.seed1 * 3 + self.seed2) % self.max_value
+        self.seed2 = (self.seed1 + self.seed2 + 33) % self.max_value
+        return float(self.seed1) / float(self.max_value)
+
+
+def scramble_old_password(password, message):
+    """Scramble for old_password"""
+    hash_pass = _hash_password_323(password)
+    hash_message = _hash_password_323(message[:SCRAMBLE_LENGTH_323])
+    hash_pass_n = struct.unpack(">LL", hash_pass)
+    hash_message_n = struct.unpack(">LL", hash_message)
+
+    rand_st = RandStruct_323(
+        hash_pass_n[0] ^ hash_message_n[0], hash_pass_n[1] ^ hash_message_n[1]
+    )
+    outbuf = io.BytesIO()
+    for _ in range_type(min(SCRAMBLE_LENGTH_323, len(message))):
+        outbuf.write(int2byte(int(rand_st.my_rnd() * 31) + 64))
+    extra = int2byte(int(rand_st.my_rnd() * 31))
+    out = outbuf.getvalue()
+    outbuf = io.BytesIO()
+    for c in out:
+        outbuf.write(int2byte(byte2int(c) ^ byte2int(extra)))
+    return outbuf.getvalue()
+
+
+def _hash_password_323(password):
+    nr = 1345345333
+    add = 7
+    nr2 = 0x12345671
+
+    # x in py3 is numbers, p27 is chars
+    for c in [byte2int(x) for x in password if x not in (' ', '\t', 32, 9)]:
+        nr ^= (((nr & 63) + add) * c) + (nr << 8) & 0xFFFFFFFF
+        nr2 = (nr2 + ((nr2 << 8) ^ nr)) & 0xFFFFFFFF
+        add = (add + c) & 0xFFFFFFFF
+
+    r1 = nr & ((1 << 31) - 1)  # kill sign bits
+    r2 = nr2 & ((1 << 31) - 1)
+    return struct.pack(">LL", r1, r2)
+
+
+# sha256_password
 
 
 def _roundtrip(conn, send_data):
@@ -48,42 +141,35 @@ def sha2_rsa_encrypt(password, salt, public_key):
     )
 
 
-class SHA256Password(object):
-
-    def __init__(self, conn):
-        self.conn = conn
-
-    def authenticate(self, pkt):
-        conn = self.conn
-
-        if conn.ssl and conn.server_capabilities & CLIENT.SSL:
-            if DEBUG:
-                print("sha256: Sending plain password")
-            data = conn.password + b'\0'
-            return _roundtrip(conn, data)
-
-        if pkt.is_auth_switch_request():
-            conn.salt = pkt.read_all()
-            if not conn.server_public_key and conn.password:
-                # Request server public key
-                if DEBUG:
-                    print("sha256: Requesting server public key")
-                pkt = _roundtrip(conn, b'\1')
-
-        if pkt.is_extra_auth_data():
-            conn.server_public_key = pkt._data[1:]
-            if DEBUG:
-                print("Received public key:\n", conn.server_public_key.decode('ascii'))
-
-        if conn.password:
-            if not conn.server_public_key:
-                raise OperationalError("Couldn't receive server's public key")
-
-            data = sha2_rsa_encrypt(conn.password, conn.salt, conn.server_public_key)
-        else:
-            data = b''
-
+def sha256_password_auth(conn, pkt):
+    if conn.ssl and conn.server_capabilities & CLIENT.SSL:
+        if DEBUG:
+            print("sha256: Sending plain password")
+        data = conn.password + b'\0'
         return _roundtrip(conn, data)
+
+    if pkt.is_auth_switch_request():
+        conn.salt = pkt.read_all()
+        if not conn.server_public_key and conn.password:
+            # Request server public key
+            if DEBUG:
+                print("sha256: Requesting server public key")
+            pkt = _roundtrip(conn, b'\1')
+
+    if pkt.is_extra_auth_data():
+        conn.server_public_key = pkt._data[1:]
+        if DEBUG:
+            print("Received public key:\n", conn.server_public_key.decode('ascii'))
+
+    if conn.password:
+        if not conn.server_public_key:
+            raise OperationalError("Couldn't receive server's public key")
+
+        data = sha2_rsa_encrypt(conn.password, conn.salt, conn.server_public_key)
+    else:
+        data = b''
+
+    return _roundtrip(conn, data)
 
 
 def caching_sha2_scramble(password, nonce):
@@ -106,68 +192,61 @@ def caching_sha2_scramble(password, nonce):
     return bytes(res)
 
 
-class CachingSHA2Password(object):
+def caching_sha2_password_auth(conn, pkt):
+    # No password fast path
+    if not conn.password:
+        return _roundtrip(conn, b'')
 
-    def __init__(self, conn):
-        self.conn = conn
+    if pkt.is_auth_switch_request():
+        # Try from fast auth
+        if DEBUG:
+            print("caching sha2: Trying fast path")
+        conn.salt = pkt.read_all()
+        scrambled = caching_sha2_scramble(conn.password, conn.salt)
+        pkt = _roundtrip(conn, scrambled)
+    # else: fast auth is tried in initial handshake
 
-    def authenticate(self, pkt):
-        conn = self.conn
+    if not pkt.is_extra_auth_data():
+        raise OperationalError(
+            "caching sha2: Unknown packet for fast auth: %s" % pkt._data[:1]
+        )
 
-        # No password fast path
-        if not conn.password:
-            return _roundtrip(conn, b'')
+    # magic numbers:
+    # 2 - request public key
+    # 3 - fast auth succeeded
+    # 4 - need full auth
 
-        if pkt.is_auth_switch_request():
-            # Try from fast auth
-            if DEBUG:
-                print("caching sha2: Trying fast path")
-            conn.salt = pkt.read_all()
-            scrambled = caching_sha2_scramble(conn.password, conn.salt)
-            pkt = _roundtrip(conn, scrambled)
-        # else: fast auth is tried in initial handshake
+    pkt.advance(1)
+    n = pkt.read_uint8()
 
+    if n == 3:
+        if DEBUG:
+            print("caching sha2: succeeded by fast path.")
+        pkt = conn._read_packet()
+        pkt.check_error()  # pkt must be OK packet
+        return pkt
+
+    if n != 4:
+        raise OperationalError("caching sha2: Unknwon result for fast auth: %s" % n)
+
+    if DEBUG:
+        print("caching sha2: Trying full auth...")
+
+    if conn.ssl and conn.server_capabilities & CLIENT.SSL:
+        if DEBUG:
+            print("caching sha2: Sending plain password via SSL")
+        return _roundtrip(conn, conn.password + b'\0')
+
+    if not conn.server_public_key:
+        pkt = _roundtrip(conn, b'\x02')  # Request public key
         if not pkt.is_extra_auth_data():
             raise OperationalError(
-                "caching sha2: Unknown packet for fast auth: %s" % pkt._data[:1]
+                "caching sha2: Unknown packet for public key: %s" % pkt._data[:1]
             )
 
-        # magic numbers:
-        # 2 - request public key
-        # 3 - fast auth succeeded
-        # 4 - need full auth
-
-        pkt.advance(1)
-        n = pkt.read_uint8()
-
-        if n == 3:
-            if DEBUG:
-                print("caching sha2: succeeded by fast path.")
-            pkt = conn._read_packet()
-            pkt.check_error()  # pkt must be OK packet
-            return pkt
-
-        if n != 4:
-            raise OperationalError("caching sha2: Unknwon result for fast auth: %s" % n)
-
+        conn.server_public_key = pkt._data[1:]
         if DEBUG:
-            print("caching sha2: Trying full auth...")
+            print(conn.server_public_key.decode('ascii'))
 
-        if conn.ssl and conn.server_capabilities & CLIENT.SSL:
-            if DEBUG:
-                print("caching sha2: Sending plain password via SSL")
-            return _roundtrip(conn, conn.password + b'\0')
-
-        if not conn.server_public_key:
-            pkt = _roundtrip(conn, b'\x02')  # Request public key
-            if not pkt.is_extra_auth_data():
-                raise OperationalError(
-                    "caching sha2: Unknown packet for public key: %s" % pkt._data[:1]
-                )
-
-            conn.server_public_key = pkt._data[1:]
-            if DEBUG:
-                print(conn.server_public_key.decode('ascii'))
-
-        data = sha2_rsa_encrypt(conn.password, conn.salt, conn.server_public_key)
-        pkt = _roundtrip(conn, data)
+    data = sha2_rsa_encrypt(conn.password, conn.salt, conn.server_public_key)
+    pkt = _roundtrip(conn, data)
