+++ pymysql/connections.py
@@ -6,7 +6,6 @@ from __future__ import print_function
 from ._compat import PY2, range_type, text_type, str_type, JYTHON, IRONPYTHON
 
 import errno
-from functools import partial
 import hashlib
 import io
 import os
@@ -88,85 +87,10 @@ TEXT_TYPES = set([
     FIELD_TYPE.VARCHAR,
     FIELD_TYPE.GEOMETRY])
 
-sha_new = partial(hashlib.new, 'sha1')
 
-DEFAULT_CHARSET = 'latin1'
+DEFAULT_CHARSET = 'latin1'  # TODO: change to utf8mb4
 
 MAX_PACKET_LEN = 2**24-1
-
-SCRAMBLE_LENGTH = 20
-
-def _scramble(password, message):
-    if not password:
-        return b''
-    if DEBUG: print('password=' + str(password))
-    stage1 = sha_new(password).digest()
-    stage2 = sha_new(stage1).digest()
-    s = sha_new()
-    s.update(message[:SCRAMBLE_LENGTH]) 
-    s.update(stage2)
-    result = s.digest()
-    return _my_crypt(result, stage1)
-
-def _my_crypt(message1, message2):
-    length = len(message1)
-    result = b''
-    for i in range_type(length):
-        x = (struct.unpack('B', message1[i:i+1])[0] ^
-             struct.unpack('B', message2[i:i+1])[0])
-        result += struct.pack('B', x)
-    return result
-
-# old_passwords support ported from libmysql/password.c
-SCRAMBLE_LENGTH_323 = 8
-
-
-class RandStruct_323(object):
-    def __init__(self, seed1, seed2):
-        self.max_value = 0x3FFFFFFF
-        self.seed1 = seed1 % self.max_value
-        self.seed2 = seed2 % self.max_value
-
-    def my_rnd(self):
-        self.seed1 = (self.seed1 * 3 + self.seed2) % self.max_value
-        self.seed2 = (self.seed1 + self.seed2 + 33) % self.max_value
-        return float(self.seed1) / float(self.max_value)
-
-
-def _scramble_323(password, message):
-    hash_pass = _hash_password_323(password)
-    hash_message = _hash_password_323(message[:SCRAMBLE_LENGTH_323])
-    hash_pass_n = struct.unpack(">LL", hash_pass)
-    hash_message_n = struct.unpack(">LL", hash_message)
-
-    rand_st = RandStruct_323(hash_pass_n[0] ^ hash_message_n[0],
-                             hash_pass_n[1] ^ hash_message_n[1])
-    outbuf = io.BytesIO()
-    for _ in range_type(min(SCRAMBLE_LENGTH_323, len(message))):
-        outbuf.write(int2byte(int(rand_st.my_rnd() * 31) + 64))
-    extra = int2byte(int(rand_st.my_rnd() * 31))
-    out = outbuf.getvalue()
-    outbuf = io.BytesIO()
-    for c in out:
-        outbuf.write(int2byte(byte2int(c) ^ byte2int(extra)))
-    return outbuf.getvalue()
-
-
-def _hash_password_323(password):
-    nr = 1345345333
-    add = 7
-    nr2 = 0x12345671
-
-    # x in py3 is numbers, p27 is chars
-    for c in [byte2int(x) for x in password if x not in (' ', '\t', 32, 9)]:
-        nr ^= (((nr & 63) + add) * c) + (nr << 8) & 0xFFFFFFFF
-        nr2 = (nr2 + ((nr2 << 8) ^ nr)) & 0xFFFFFFFF
-        add = (add + c) & 0xFFFFFFFF
-
-    r1 = nr & ((1 << 31) - 1)  # kill sign bits
-    r2 = nr2 & ((1 << 31) - 1)
-    return struct.pack(">LL", r1, r2)
-
 
 def pack_int24(n):
     return struct.pack('<I', n)[:3]
@@ -940,14 +864,18 @@ class Connection(object):
                 auth_packet = self._process_auth(plugin_name, auth_packet)
             else:
                 # send legacy handshake
-                data = _scramble_323(self.password, self.salt) + b'\0'
+                data = _auth.scramble_old_password(self.password, self.salt) + b'\0'
                 self.write_packet(data)
                 auth_packet = self._read_packet()
         elif auth_packet.is_extra_auth_data():
             print("received extra data")
             # https://dev.mysql.com/doc/internals/en/successful-authentication.html
-            handler = self._get_auth_plugin_handler(self._auth_plugin_name)
-            handler.authenticate(auth_packet)
+            if self._auth_plugin_name == b"caching_sha2_password":
+                auth_packet = _auth.caching_sha2_password_auth(self, auth_packet)
+            elif self._auth_plugin_name == b"sha256_password":
+                auth_packet = _auth.sha256_password_auth(self, auth_packet)
+            else:
+                raise OperationalError("Received extra packet for auth method %r", self._auth_plugin_name)
 
         if DEBUG: print("Succeed to auth")
 
@@ -960,12 +888,14 @@ class Connection(object):
                 if plugin_name != b'dialog':
                     raise err.OperationalError(2059, "Authentication plugin '%s'" \
                               " not loaded: - %r missing authenticate method" % (plugin_name, plugin_class))
-        if plugin_name == b"mysql_native_password":
-            # https://dev.mysql.com/doc/internals/en/secure-password-authentication.html#packet-Authentication::Native41
-            data = _scramble(self.password, auth_packet.read_all())
+        if plugin_name == b"caching_sha2_password":
+            data = _auth.caching_sha2_password_auth(self, auth_packet)
+        elif plugin_name == b"sha256_password":
+            data = _auth.sha256_password_auth(self, auth_packet)
+        elif plugin_name == b"mysql_native_password":
+            data = _auth.scramble_native_password(self.password, auth_packet.read_all())
         elif plugin_name == b"mysql_old_password":
-            # https://dev.mysql.com/doc/internals/en/old-password-authentication.html
-            data = _scramble_323(self.password, auth_packet.read_all()) + b'\0'
+            data = _auth.scramble_old_password(self.password, auth_packet.read_all()) + b'\0'
         elif plugin_name == b"mysql_clear_password":
             # https://dev.mysql.com/doc/internals/en/clear-text-authentication.html
             data = self.password + b'\0'
